apps/api/pyproject.toml
[project]
name = "jumlaos-api"
version = "0.1.0"
description = "JumlaOS API — modular monolith for Moroccan wholesalers"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "Proprietary" }
authors = [{ name = "JumlaOS Team" }]
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "sqlalchemy[asyncio]>=2.0.32",
  "alembic>=1.13",
  "asyncpg>=0.29",
  "psycopg[binary]>=3.2",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
  "redis>=5.0",
  "procrastinate>=2.12",
  "orjson>=3.10",
  "python-multipart>=0.0.12",
  "email-validator>=2.2",
  "phonenumbers>=8.13",
  "rapidfuzz>=3.10",
  "structlog>=24.4",
  "sentry-sdk[fastapi]>=2.14",
  "tenacity>=9.0",
  "boto3>=1.35",
  "itsdangerous>=2.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7",
  "mypy>=1.13",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "pytest-cov>=5.0",
  "pytest-socket>=0.7",
  "hypothesis>=6.115",
  "respx>=0.21",
  "freezegun>=1.5",
  "types-python-jose",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jumlaos"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
  "E", "F", "W",      # pycodestyle / pyflakes
  "I",                 # isort
  "B",                 # bugbear
  "UP",                # pyupgrade
  "SIM",               # simplify
  "RUF",               # ruff
  "ASYNC",             # async correctness
  "S",                 # bandit-like
  "T20",               # no print
  "TID",               # clean imports
  "PIE", "PT", "PL",
]
ignore = [
  "E501",              # line length handled by formatter
  "PLR0913",           # function call arg count
  "PLR2004",           # magic numbers (too noisy in tests)
  "S101",              # allow assert in tests
  "S105", "S106",      # false positives for fixture "passwords"
  "B008",              # FastAPI Depends() in defaults is intentional
  "UP042",             # str+Enum kept for SQLAlchemy Enum compat
  "PLC0415",           # nested imports tolerated to break dep cycles
  "ASYNC110",          # in worker entrypoint, a sleep loop is fine
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "PLR", "PT"]
"alembic/**" = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["jumlaos"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_unused_ignores = true
disallow_untyped_decorators = false     # FastAPI/SQLAlchemy decorators are fine
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "procrastinate.*",
  "boto3.*",
  "botocore.*",
  "passlib.*",
  "jose.*",
  "rapidfuzz.*",
  "phonenumbers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --disable-socket --allow-unix-socket --allow-hosts=127.0.0.1,::1,localhost"
filterwarnings = [
  "error",
  "ignore::DeprecationWarning:passlib.*",
]

[tool.coverage.run]
branch = true
source = ["jumlaos"]

[tool.coverage.report]
skip_covered = true
show_missing = true
fail_under = 60
